development/requirements.txt
numpy==1.19.5
tensorflow-gpu
pillow
opencv-python
gdown
flask